+++ doc/whats_new.rst
@@ -467,9 +467,9 @@ Bug fixes
       (`#7350 <https://github.com/scikit-learn/scikit-learn/pull/7350/>`_)
       By `Russell Smith`_.
 
-    - Fix regression in :func:`metrics.silhouette_score` in which clusters of
-      size 1 were incorrectly scored. By `Ramana Subramanyam
-      <https://github.com/Sentient07>`_.
+    - Fix bug in :func:`metrics.silhouette_score` in which clusters of
+      size 1 were incorrectly scored. They should get a score of 0.
+      By `Joel Nothman`_.
 
 
 API changes summary